Add polySymmetryData node caching
polySymmetryData nodes are cached in memory, which will allow other tools
in this plugin to operate on a mesh without having to get bogged down with
keeping track of the nodes and passing the right node to a command.

The polySymmetry command -query flag can be used to fetch the polySymmetryData
node associated with a mesh and pack the data on that node into a JSON
object, allowing Python scripts to leverage the symmetry data.